Update README.md
Bennett Gray V, opensea and Rachael Browner, and her associates, are liable for all damages, personal, and public, or If blockchain or any copies are taken without permission. Bennett Gray V is not her associate, and Rachael Browner and her associates are liable to spend the rest of there lives, in federal prison.